<작업일시> 202501141430 <작업자> 강민서
<작업내용>
스키마 및 테스트케이스 정리
