+++ src/main/java/home/project/config/redis/PageImplDeserializerModule.java
@@ -1,0 +1,10 @@
+package home.project.config.redis;
+
+import com.fasterxml.jackson.databind.module.SimpleModule;
+import org.springframework.data.domain.PageImpl;
+
+public class PageImplDeserializerModule extends SimpleModule {
+    public PageImplDeserializerModule() {
+        addDeserializer(PageImpl.class, new PageImplDeserializer());
+    }
+}
